[MOS-807] Unpacking update package
* Added unpacking and validating of
update package
* Removed 'utils-bootconfig'
* Added 'tar' module
* Minor compilation issues fixed

+++ module-utils/time/time/time_conversion.cpp
@@ -270,7 +270,11 @@ namespace utils::time
         seconds   = (this->duration % secondsInHour) % secondsInMinute;
 
         if (verboseConversion) {
-            LOG_DEBUG("duration %" PRIu64 " - %lu hours %lu minutes %lu seconds", duration, hours, minutes, seconds);
+            LOG_DEBUG("duration %" PRIu64 " - %lu hours %lu minutes %lu seconds",
+                      static_cast<std::uint64_t>(duration),
+                      hours,
+                      minutes,
+                      seconds);
         }
     }
